+++ lib/ts/recipe/dashboard/api/implementation.ts
@@ -1,0 +1,63 @@
+/* Copyright (c) 2022, VRAI Labs and/or its affiliates. All rights reserved.
+ *
+ * This software is licensed under the Apache License, Version 2.0 (the
+ * "License") as published by the Apache Software Foundation.
+ *
+ * You may not use this file except in compliance with the License. You may
+ * obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ */
+
+import NormalisedURLDomain from "../../../normalisedURLDomain";
+import NormalisedURLPath from "../../../normalisedURLPath";
+import SuperTokens from "../../../supertokens";
+import { DASHBOARD_API } from "../constants";
+import { APIInterface } from "../types";
+
+export default function getAPIImplementation(): APIInterface {
+    return {
+        dashboardGET: async function (input) {
+            const bundleBasePathString = await input.options.recipeImplementation.getDashboardBundleLocation({
+                userContext: input.userContext,
+            });
+
+            const bundleDomain =
+                new NormalisedURLDomain(bundleBasePathString).getAsStringDangerous() +
+                new NormalisedURLPath(bundleBasePathString).getAsStringDangerous();
+
+            let connectionURI: string = "";
+            const superTokensInstance = SuperTokens.getInstanceOrThrowError();
+
+            if (superTokensInstance.supertokens !== undefined) {
+                connectionURI = superTokensInstance.supertokens.connectionURI;
+            }
+
+            return `
+            <html>
+                <head>
+                    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+                    <script>
+                        window.staticBasePath = "${bundleDomain}/static"
+                        window.dashboardAppPath = "${input.options.appInfo.apiBasePath
+                            .appendPath(new NormalisedURLPath(DASHBOARD_API))
+                            .getAsStringDangerous()}"
+                        window.connectionURI = "${connectionURI}"
+                    </script>
+                    <script defer src="${bundleDomain}/static/js/bundle.js"></script></head>
+                    <link href="${bundleDomain}/static/css/main.css" rel="stylesheet" type="text/css">
+                    <link rel="icon" type="image/x-icon" href="${bundleDomain}/static/media/favicon.ico">
+                </head>
+                <body>
+                    <noscript>You need to enable JavaScript to run this app.</noscript>
+                    <div id="root"></div>
+                </body>
+            </html>
+            `;
+        },
+    };
+}