+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Adds:
+
+-   Adds a Dashboard recipe
+
 ## [11.1.2] - 2022-08-25
 
 ### Type fix: